requirements.txt
openai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pinecone
sentence-transformers>=2.2.0
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
networkx
